requirements.txt
altair==5.5.0
attrs==25.3.0
biopython==1.85
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dotenv==0.9.9
gitdb==4.0.12
gitpython==3.1.45
idna==3.10
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
narwhals==2.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
